souoc

+++ so uoc.c
@@ -1,15 +1,15 @@
 #include <stdio.h>
 
 int main(){
-	int n;
+	long long n;
 	scanf("%d",&n);
-	int i;
-	int so_uoc = 0;
+	long long i;
+	long long so_uoc = 0;
 	for(i = 1; i <= n; i++){
 		if(n % i == 0){
 			so_uoc ++;
 		}
 	}
-	printf("%d",so_uoc);
+	printf("%lld",so_uoc);
 	return 0;
 }
